requirements.txt
Flask==2.3.3
Authlib==1.2.1
python-dotenv==1.0.0
requests==2.31.0
Flask-SocketIO==5.3.6
eventlet==0.35.2
redis==5.0.1
psutil==5.9.6